requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.5
    # via openai
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via httpcore
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
black==23.7.0
    # via -r requirements.in
certifi==2023.7.22
    # via
    #   httpcore
    #   requests
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
click==8.1.6
    # via
    #   black
    #   typer
dnspython==2.4.0
    # via pinecone-client
exceptiongroup==1.1.2
    # via anyio
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
httpcore==0.17.3
    # via dnspython
idna==3.4
    # via
    #   anyio
    #   requests
    #   yarl
isort==5.12.0
    # via -r requirements.in
loguru==0.7.0
    # via pinecone-client
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
numpy==1.25.1
    # via pinecone-client
openai==0.27.8
    # via -r requirements.in
packaging==23.1
    # via black
pathspec==0.11.1
    # via black
pinecone-client==2.2.2
    # via -r requirements.in
platformdirs==3.9.1
    # via black
pydantic==2.0.3
    # via
    #   -r requirements.in
    #   pydantic-settings
pydantic-core==2.3.0
    # via pydantic
pydantic-settings==2.0.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via pinecone-client
python-dotenv==1.0.0
    # via pydantic-settings
pyyaml==6.0.1
    # via pinecone-client
regex==2023.6.3
    # via tiktoken
requests==2.31.0
    # via
    #   openai
    #   pinecone-client
    #   tiktoken
ruff==0.0.280
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   dnspython
    #   httpcore
tiktoken==0.4.0
    # via -r requirements.in
toml==0.10.2
    # via -r requirements.in
tomli==2.0.1
    # via black
tqdm==4.65.0
    # via
    #   openai
    #   pinecone-client
typer==0.9.0
    # via -r requirements.in
typing-extensions==4.7.1
    # via
    #   pinecone-client
    #   pydantic
    #   pydantic-core
    #   typer
urllib3==2.0.4
    # via
    #   pinecone-client
    #   requests
yarl==1.9.2
    # via aiohttp
